feat(web): convert to single page app
