--- packages/web/overview.html
@@ -1,60 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Kingdom Builder - Overview</title>
-    <style>
-      body {
-        max-width: 800px;
-        margin: 2rem auto;
-        font-family: sans-serif;
-        line-height: 1.5;
-        padding: 0 1rem;
-      }
-      h1 {
-        text-align: center;
-      }
-    </style>
-  </head>
-  <body>
-    <h1>Game Overview</h1>
-    <p>
-      Kingdom Builder is a turn-based duel where you grow your realm and attempt
-      to outmaneuver your rival. Protect your castle, expand your lands, and
-      manage your resources to prevail.
-    </p>
-    <h2>Goal</h2>
-    <p>
-      Outlast or conquer your opponent. A game can end when a castle falls, when
-      a player can no longer sustain their realm, or when the final round ends
-      with one ruler holding the advantage.
-    </p>
-    <h2>Phases</h2>
-    <p>Each turn flows through three phases:</p>
-    <ul>
-      <li>
-        <strong>Development</strong> – your realm produces resources and
-        triggered effects take place.
-      </li>
-      <li>
-        <strong>Upkeep</strong> – maintain your population and resolve ongoing
-        effects.
-      </li>
-      <li>
-        <strong>Main</strong> – both players secretly choose actions and then
-        resolve them in turn.
-      </li>
-    </ul>
-    <h2>Core Mechanics</h2>
-    <p>
-      Actions may require resources or other prerequisites and grant various
-      effects, such as gaining resources, building structures, developing land,
-      or hindering your opponent. Buildings provide passive bonuses, while land
-      around your castle holds developments that yield benefits. Resources like
-      Gold, Action Points, Happiness, and Castle Health are spent and gained
-      throughout the game.
-    </p>
-    <p><a href="index.html">Back to Start</a></p>
-  </body>
-</html>
