+++ packages/web/tests/Game.render.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from 'vitest';
 import { renderToString } from 'react-dom/server';
 import React from 'react';
-import App from '../src/App';
+import Game from '../src/Game';
 
 // Render the application with the real engine to ensure that
 // dynamic action effects (e.g. build with "$id") don't crash
@@ -10,8 +10,8 @@ vi.mock('@kingdom-builder/engine', async () => {
   // Re-export the actual engine source since vitest doesn't resolve the monorepo alias
   return await import('../../engine/src');
 });
-describe('<App /> integration', () => {
+describe('<Game /> integration', () => {
   it('renders without crashing', () => {
-    expect(() => renderToString(<App />)).not.toThrow();
+    expect(() => renderToString(<Game />)).not.toThrow();
   });
 });
